YouTube Video Downloader/requirements.txt
pytube
Pillow
requests


# pip install -r requirements.txt
